# Conflicts:
#	CHANGELOG.md
#	composer.lock
#	includes/ClicShopping/External/vendor/composer/installed.json
#	includes/ClicShopping/External/vendor/composer/installed.php
#	includes/ClicShopping/Sites/Shop/RewriteUrl.php
#	includes/ClicShopping/version.json
#	install/Db/clicshopping.sql
#	install/Db/clicshopping_en.sql

+++ includes/ClicShopping/Apps/Catalog/Manufacturers/Sites/ClicShoppingAdmin/Pages/Home/Actions/Manufacturers/DeleteAll.php
@@ -21,7 +21,7 @@ class DeleteAll extends \ClicShopping\OM\PagesActionsAbstract
     $this->app = Registry::get('Manufacturers');
     $this->Hooks = Registry::get('Hooks');
 
-    if (isset($_POST['selected'])) {
+    if (isset($_POST['selected']) && is_array($_POST['selected']) && isset($_POST['DeleteAll'])) {
       foreach ($_POST['selected'] as $id) {
 
         $this->app->db->delete('manufacturers', ['manufacturers_id' => (int)$id]);
@@ -38,7 +38,7 @@ class DeleteAll extends \ClicShopping\OM\PagesActionsAbstract
 
         $Qupdate->execute();
 
-        $this->Hooks->call('Manufacturers', 'Delete');
+        $this->Hooks->call('Manufacturers', 'DeleteAll');
       }
     }
